Add encryption mode support
This change adds the ability to have different encryption modes.
Currently only AES-CBC is supported, but this will make it simple to
add new modes such as secretbox.  The mode is configurable via a new
homonymous command line flag.  The mode specification is stored as a
prefix with the encrypted data.  It is validated during decryption.

The format is:

:ck:<mode_name>:<mode_version>:

Thus for AES-CBC the value is:

:ck:aescbc:v1:

The mode version is reserved in case we need to make any backwards
incompatible changes to how the encrypted data is stored.

Signed-off-by: Monis Khan <mkhan@redhat.com>